requirements.txt
absl-py==0.9.0
astunparse==1.6.3
attrs==19.3.0
audio-metadata==0.11.1
bidict==0.19.0
bitstruct==8.10.0
boto3==1.13.19
botocore==1.16.19
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cycler==0.10.0
docutils==0.15.2
Flask==1.1.2
gast==0.3.3
gcloud==0.17.0
google-api-core==1.17.0
google-auth==1.15.0
google-auth-oauthlib==0.4.1
google-cloud-language==1.3.0
google-cloud-speech==1.3.2
google-pasta==0.2.0
googleapis-common-protos==1.51.0
grpcio==1.29.0
gunicorn==20.0.4
gurunudi==1.3.8
h5py==2.10.0
httplib2==0.18.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
jws==0.1.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
more-itertools==8.3.0
numpy==1.18.4
oauth2client==3.0.0
oauthlib==3.1.0
opt-einsum==3.2.1
pendulum==2.0.4
Pillow==7.1.2
playsound==1.2.2
pprintpp==0.4.0
protobuf==3.12.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.4.3
pyparsing==2.4.7
Pyrebase==3.0.27
PySoundFile==0.9.0.post1
python-dateutil==2.8.1
python-jwt==2.0.1
pytz==2020.1
pytzdata==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.7.0
rsa==4.0
scipy==1.4.1
six==1.15.0
sox==1.3.7
SpeechRecognition==3.8.1
tbm-utils==2.6.0
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
urllib3==1.25.9
wavio==0.0.4
Werkzeug==1.0.1
wrapt==1.12.1
